refactor(studio): extract DecisionsSection from decisions route
